refactor: rename testing/api to testing/controller
