+++ src/controller/event/handler/webhook/scan/scan_test.go
@@ -31,8 +31,8 @@ import (
 	"github.com/goharbor/harbor/src/pkg/scan/dao/scan"
 	"github.com/goharbor/harbor/src/pkg/scan/report"
 	v1 "github.com/goharbor/harbor/src/pkg/scan/rest/v1"
-	artifacttesting "github.com/goharbor/harbor/src/testing/api/artifact"
-	scantesting "github.com/goharbor/harbor/src/testing/api/scan"
+	artifacttesting "github.com/goharbor/harbor/src/testing/controller/artifact"
+	scantesting "github.com/goharbor/harbor/src/testing/controller/scan"
 	"github.com/goharbor/harbor/src/testing/mock"
 	notificationtesting "github.com/goharbor/harbor/src/testing/pkg/notification"
 	"github.com/stretchr/testify/require"
