+++ src/controller/scan/base_controller_test.go
@@ -36,8 +36,8 @@ import (
 	"github.com/goharbor/harbor/src/pkg/scan/dao/scanner"
 	v1 "github.com/goharbor/harbor/src/pkg/scan/rest/v1"
 	"github.com/goharbor/harbor/src/pkg/scan/vuln"
-	artifacttesting "github.com/goharbor/harbor/src/testing/api/artifact"
-	scannertesting "github.com/goharbor/harbor/src/testing/api/scanner"
+	artifacttesting "github.com/goharbor/harbor/src/testing/controller/artifact"
+	scannertesting "github.com/goharbor/harbor/src/testing/controller/scanner"
 	mocktesting "github.com/goharbor/harbor/src/testing/mock"
 	reporttesting "github.com/goharbor/harbor/src/testing/pkg/scan/report"
 	"github.com/stretchr/testify/assert"
