+++ src/server/middleware/quota/quota_test.go
@@ -26,10 +26,10 @@ import (
 	"github.com/goharbor/harbor/src/controller/project"
 	"github.com/goharbor/harbor/src/controller/quota"
 	"github.com/goharbor/harbor/src/pkg/types"
-	artifacttesting "github.com/goharbor/harbor/src/testing/api/artifact"
-	blobtesting "github.com/goharbor/harbor/src/testing/api/blob"
-	projecttesting "github.com/goharbor/harbor/src/testing/api/project"
-	quotatesting "github.com/goharbor/harbor/src/testing/api/quota"
+	artifacttesting "github.com/goharbor/harbor/src/testing/controller/artifact"
+	blobtesting "github.com/goharbor/harbor/src/testing/controller/blob"
+	projecttesting "github.com/goharbor/harbor/src/testing/controller/project"
+	quotatesting "github.com/goharbor/harbor/src/testing/controller/quota"
 	"github.com/goharbor/harbor/src/testing/mock"
 	"github.com/stretchr/testify/suite"
 )
