+++ src/server/registry/catalog_test.go
@@ -18,7 +18,7 @@ import (
 	"encoding/json"
 	"github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/controller/repository"
-	repotesting "github.com/goharbor/harbor/src/testing/api/repository"
+	repotesting "github.com/goharbor/harbor/src/testing/controller/repository"
 	"github.com/stretchr/testify/suite"
 	"net/http"
 	"net/http/httptest"
