+++ src/server/registry/tag_test.go
@@ -20,8 +20,8 @@ import (
 	"github.com/goharbor/harbor/src/controller/repository"
 	"github.com/goharbor/harbor/src/controller/tag"
 	model_tag "github.com/goharbor/harbor/src/pkg/tag/model/tag"
-	repotesting "github.com/goharbor/harbor/src/testing/api/repository"
-	tagtesting "github.com/goharbor/harbor/src/testing/api/tag"
+	repotesting "github.com/goharbor/harbor/src/testing/controller/repository"
+	tagtesting "github.com/goharbor/harbor/src/testing/controller/tag"
 	"github.com/stretchr/testify/suite"
 	"net/http"
 	"net/http/httptest"
