+++ src/server/v2.0/handler/assembler/vul_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 
 	"github.com/goharbor/harbor/src/server/v2.0/handler/model"
-	"github.com/goharbor/harbor/src/testing/api/scan"
+	"github.com/goharbor/harbor/src/testing/controller/scan"
 	"github.com/goharbor/harbor/src/testing/mock"
 	"github.com/stretchr/testify/suite"
 )
