+++ src/testing/controller/scan/checker.go
